+++ BigQuery_procedures/merge_d_Regiao_Turistica.sql
@@ -1,0 +1,34 @@
+CREATE OR REPLACE PROCEDURE `tccfacens2024-435322.TCC_Turismo.merge_d_Regiao_Turistica`()
+BEGIN
+
+
+MERGE `tccfacens2024-435322.TCC_Turismo.d_Regiao_Turistica` tgt
+USING (SELECT 
+  NO_MUNICIPIO,
+  REGIAO_TURISTICA,
+  CATEGORIA_TURISMO
+FROM `tccfacens2024-435322.TCC_Turismo_staging.d_Regiao_Turistica` ) as src
+ON
+  tgt.NO_MUNICIPIO  = src.NO_MUNICIPIO and 
+  tgt.REGIAO_TURISTICA  = src.REGIAO_TURISTICA and 
+  tgt.CATEGORIA_TURISMO  = src.CATEGORIA_TURISMO 
+
+WHEN MATCHED THEN
+  UPDATE SET 
+    tgt.NO_MUNICIPIO  = src.NO_MUNICIPIO,
+    tgt.REGIAO_TURISTICA  = src.REGIAO_TURISTICA,
+    tgt.CATEGORIA_TURISMO  = src.CATEGORIA_TURISMO 
+  
+WHEN NOT MATCHED BY TARGET THEN
+  INSERT (
+    NO_MUNICIPIO ,
+    REGIAO_TURISTICA ,
+    CATEGORIA_TURISMO  )
+  VALUES (
+    src.NO_MUNICIPIO ,
+    src.REGIAO_TURISTICA ,
+    src.CATEGORIA_TURISMO 
+    );
+    
+DELETE FROM `tccfacens2024-435322.TCC_Turismo_staging.d_Regiao_Turistica` WHERE 1 = 1;
+END;
