--- CloudFunction/requirements.txt
@@ -1,3 +1,0 @@
-functions-framework
-google-cloud-bigquery==3.3.1
-numpy==1.26.4
